% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrpc.R
\name{encodePlot}
\alias{encodePlot}
\title{Renders a plot as a base64-encoded png}
\usage{
encodePlot(width, height, plotFn)
}
\arguments{
\item{width}{Width of the plot in pixels}

\item{height}{Height of the plot in pixels}

\item{plotFn}{Function to call to perform the plot}
}
\description{
Renders a plot as a base64-encoded png
}
