% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrpc.R
\name{encodePlotAsPng}
\alias{encodePlotAsPng}
\title{Renders a plot as a base64-encoded PNG}
\usage{
encodePlotAsPng(width, height, plotFn)
}
\arguments{
\item{width}{Width of the plot in pixels}

\item{height}{Height of the plot in pixels}

\item{plotFn}{Function to call to perform the plot}
}
\description{
The result can be set as the \code{src} attribute of an \code{img} element in HTML.
}
\seealso{
\code{\link[=rrpcServer]{rrpcServer()}}
}
