% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrpc.R
\name{rrpc}
\alias{rrpc}
\title{Returns the function that sets up the WebSocket}
\usage{
rrpc(interface)
}
\arguments{
\item{interface}{A list of named functions that will be able to be called
over the WebSocket. Each function can have parameters that will be fed
from the params object passed in, the key in the params object matching
the parameter name.}
}
\description{
If calling the \code{httpuv::startServer} function manually, \code{rrpc}
returns the function to provde as the \code{onWSOpen} parameter.
}
\examples{
httpuv::startServer(host='127.0.0.1', port=8888, app=list(
  onWSOpen=rrpc(list(
    add=function(a,b) {
      a + b
    },
    multiply=function(a,b) {
      a * b
    }
  ))
))
}
\seealso{
\code{\link[httpuv:startServer]{httpuv::startServer()}}
}
